Camera Bug Fix
If permissions for camera is denied, app will request for permissions again if the user clicks on the rescan button again. If the user allows camera permissions, the app will not ask again because permissions are allowed.
I also removed a type cast because Android Studio said it was redundant.
-Littlesnowman88