Added the functionality to have a list of events
When pressing the QR scan button at the top of the screen, it now drops down a list of recently used events. There is a 'Scan QR' selection that will allow the user to scan a new QR code.

The list allows there to be 5 events, and will remove the least recently used event when loading a new event if there are already 5 events.

As of right now, the display of the event names are not arranged properly in the QR event list drop down. This is still being worked on.

JSONs created before summer of 2018 will crash the settings activity unless the remote zone and custom zone options are built into the JSONs.

The app still has yet to save the past QR events upon closing.

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>