Fixed the QR drop down menu
The drop down menu now has the events in the correct order.  This issue was not in the list but in the dataconnection.
Added runnable variable to Dataconnect so it could run resetScannedEventsAdapter in post execute
Event name will need to be added in place of timezone, but data is not currently avaible through the web app.

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>
Co-Authored-By: D. Shade <donaldrshade@users.noreply.github.com>