Added a 'no event' state. Also made the settings compatible with pre-webapp update JSON files
Previously, attempting to go to the settings would result in the app crashing unless using a post web app update JSON. A no event state was added for when there are no events scanned.  Older JSON files will now have a default settings page that will not crash the app.

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>