The app now uses the event name instead of the timezone in the dropdown list. It also lists an event as "No Name" if no name is provided in the JSON
The scanned events drop down previously used the timezone as the event name because the JSON did not provide an event name. Now that the JSONs provide an event name, the app displays that instead of the timezone.

If the JSON does not have an event name or the name is all spaces, the event will be shown as "No Name".

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>