Auto Update Setting Fixed
I realized my "auto" algorithm for the Auto Updater's setRefreshFrequency was incorrect, so I fixed a logic error with a helper function.
Also, I updated the "auto" algorithm so that (when on cell data) the algorithm sets its refresh frequency when EITHER an hour has passed OR the refresh button has been pressed. The latter function is the newer update.
The helper class is not completely foolproof from abuse. Currently it calls Auto Updater's setAutoRefresh() if it is called by MainActivity. Other things can pass in MainActivity.class and pretend to be Main Activity, though, so this is a problem.