Menu Text Contrast Refactor
reduced duplicate code with the creation of the ColorContrastHelper.
Also made LocationInfo final, since the time zone info class will be final when the updated settings recycler view for custom time zones is finished.
IMPORTANT: if multiple event locations are implemented, the LocationInfo class's static status will need to be examined. I don't know at this point if it will need to stay static or not.