Organized the time zones for the setting UI update, UI is not updated yet
TimeZoneInfo will now return the time zones in a more organized way. This is so that the settings time zone recycle view will not be a long list of time zones, but rather a list of continents that can be pressed to bring up a list of their time zones.

The UI has not yet been updated to use this.

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>