Settings UI Refactor
For Custom Time Zone setting (and the structure is set up for multiple locations, too), the UI is now easier for the users. Originally, the users had to scroll through one big list with all available time zone IDs. tfmoo and I parsed the list of zone IDs, sorting them by continent. When the user clicks on the continent, he or she then selects a specific zone within the continent.