Notifications Changed Update
The event app will no longer send a notificaiton for already-notified notifications.
The event app will notify the user when a new notification is provided.
The event app will notify the user when an existing notification is changed.
The refresh button now checks for a notification update, thanks to a logic fix in DataConnection's LoadGeneralInfo.