Notifications Update Fix
Upon switching events, the notifications would all be marked as "new". I changed the app's functionality so that changing events marks the notifications as "old"--so the user doesn't get bombarded with notifications that he/she can see on the screen anyways.
I discovered that loading a QR code triggers the device's default notification ringtone. I tweaked the app so that choosing a DIFFERENT QR code triggers the device's default notification ringtone. Choosing the current QR code simply behaves as a refresh.