Schedule Day Range Fix
tfmoo and I found an instance in which schedule infinitely created "tomorrow" because the JSON days were out of chronological order.
Now, the schedule should not infinitely create days either forward or backwards, and it should always start with the earliest day (even when the first JSON schedule item isn't the earliest day).