About page is no longer an 'Information Page'
The about page is now an about page instead of an information page. This will allow for information pages to have titles like "About the Event". Previously, an information page with a title beginning with "About" would also put the information page's information in the About page. This has been fixed.