Information Page Bug Fixes
2 fixes here.
When a new information page is created but not saved, the icon would save as a blank string and crash the app, requiring a reinstall. I put a safety ternary statement to fix this (as well as some readability refactoring).
If information pages had names "About the Event" and "About the Event 2", "About the Event" would load contents from both pages. We fixed this problem (or at least band-aided it) with a "LIMIT" statement.

Co-Authored-By: tfmoo <tfmoo@users.noreply.github.com>