Mysterious Auto-Update bug fixed?
I am pushing this before thorougher testing of the mysterious bug that tfmoo and I have encountered with a web app json. I kept seeing a null pointer exception in the logcat, so I set out to fix that. The error had to do with notification titles, and the mysterious bug also had to do (somewhat) with notification titles, so hopefully this patch will have fixed the bug. Next up is un-hardcoding more strings, the right way: with Context, not activity.