Refresh Bug Fixes
Two small bug fixes in this commit.
First, I placed a try/catch block around the creation of a spinner. I think I know why it breaks (and documented about that in the code), but it was stopping refresh from properly loading config changes correctly.
I made a method private and non-static because it no longer needed to be public or static. I think that's because the About page is now its own page.
Third, when the app loaded up from a fresh install, my "boolean isSameURL" threw an error becuase old_qrAddress was null. We swapped the position of qrAddress and old_qrAddress because qrAddress is always provided, and old_qrAddress--if it's null--should be different anyway!