Expired Event Scanning (Partial Fix)
If an event expires, the app won't load the event JSON. fine.
But then if I changed the expiration date, our version number implementation prevented the app from loading the recently un-expired event. I changed this so that, if a qr is directly scanned, the app will always load the event as if it were new.