Database Upgrade Compatibility
Database can now be upgraded when the app is updated. This will add the Scanned Events table, About Page table, and JSON Version Number table to the already existing database if the database version is out of date. It will also add a Nav ID column to the navigation titles table. This has not yet been tested.

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>