Database Upgrade fix attempt 1
If an old version of an app has an event loaded and then the user upgrades the app, the user would have to rescan a qr in order to get the information back. This is the first attempt to fix that problem, by preventing the app from going into MainActivity's handleNoScannedEvent().