Database Upgrade fix attempt 3
I had to add a stipulation about version numbers.
If a version number is 0,0, then then app should always update both the config and notifications when updating an app.
This is for backwards compatibility with old JSONs that may not have version numbers.
I am wondering if I will need to make it so that version number does not have to be the first item in the database's "general" section... so that if future ordering changes or someone decides to hand-type a JSON, the app can be flexible and still work.