Scanned Events and No Connection fixes
I found problems with the event app when I started using it outside of my local area network, where the JSONs were stored. So, tfmoo and I have begun debugging the app when it cannot connect to its qr-code.
For this update, tfmoo and I added a new logo and some information indicating when a scanned event tried to scan and failed to connect. We also fixed several bugs in the process.
We redesigned our main activity runnable in DataConnection. Or rather, we removed it and did message passing the way we are supposed to do so in Android: through intents. (One would think I would learn...)
More fixes to come, including:
Schedule getting stuck in an infinite loop when not connected to its QR code.
Proper handling of deleting the current event from scanned events.
Going back to the no-event state when all events have been deleted.
Also, the about-page bug is back :\
-Littlesnowman88

Co-Authored-By: tfmoo <tfmoo@users.noreply.github.com>