Deleting an Event Changes
Previously, deleting the current event would keep the event loaded until a different event is loaded. Now when deleting the current event, it will load the next most recent event. This will soon be changed so that the app loads a blank event instead after deleting the current event.

Also, when all events are deleted, the 'no event' state is loaded.