Backwards Compatibility Crash Fix
Because I tried to allow old JSONs to refresh, I ended up double-loading the event data for the old JSONs. This wasn't a problem until the schedule tried to load. The double-loading of schedule activities caused an infinite loop because the schedule is not designed for handling overlapping events!
In short, schedule infinite loop is fixed, but old JSON data CANNOT be refreshed without manually adding and changing version numbers.