Current Schedule Event now Yellow
When AutoUpdater checks for new notifications, the schedule recreates. This allows the schedule to update according to the current time. If the current time overlaps with the time of an activity in the schedule, that schedule item will have a yellow background. This also happens during times where there are no scheduled items.

This is so the user is more easily able to identify what shedule activity is currently ongoing.

Co-Authored-By: Andrew Fulling <littlesnowman88@users.noreply.github.com>