Multiple Event Location Handling, part 1
tfmoo and I made dummy data for the event App to process multiple event locations.
A merge went wierd for me and so I may have lost some documentation... :(
tfmoo and I plan on having the app make the user choose an event location when a "new" data connection is created, the user will be prompted to select an event location (if multiple exist). The user can then select an event location, or hit cancel and thus default to the first event location.
Once JSONs can support multiple event locations, this branch will need to be finished. Make sure to keep backwards compatibility!
-Littlesnowman88

Co-Authored-By: tfmoo <tfmoo@users.noreply.github.com>