Initial Location Selection Activity
When an event has multiple event locations, the app will now ask the user which location they are attending when the event is initially loaded. This is done through the use of a recyclerview with radio buttons indicating the user's selection. By default, the first location listed has the checked radio button. Choosing a location will adjust the schedule, as well as the preference summary in the settings.

The recycler view also has a 'Remote Location' option. For now, this is the device's location. This may be made toggleable in the future in LocationSelectionRecyclerView.java.

Other changes:
Previously with multiple event locations, schedule items did not have a reference location in order to properly adjust for timezones. When there are multiple event locations, the host will need to choose a main location. Right now, the app uses the first event location listed in the JSON. It may be useful for the web app to have a main location at which the schedule items will be set at.