Android Oreo (8.0) Notifications Fix
Previously, devices running Android Oreo would not receive notifications outside of the app. This is because Android Oreo requires notifications to have a notification channel. If the device's version is Oreo or newer, a notification channel will now be created.

Additionally, clicking on a notification will now clear the notification from the device's notification panel.